Deletion Bug fixed

+++ app/api/attachment/clear/route.ts
@@ -24,7 +24,9 @@ export async function GET(request: NextRequest) {
         message: "User Not Found",
       });
     }
-    await fs.promises.unlink(filepath);
+    if (fs.existsSync(filepath)) {
+      await fs.promises.unlink(filepath);
+    }
     return Response.json({
       success: true,
       messages: "Attachment Deleted in Buffer",
